Switch to randomized choice of triager (excluding the bug creator)
Today's list of bugs included enough of the triage team that the
shuffle-based approach failed on every try.

Here we replace it with a simpler approach: randomly pick a triager,
excluding the bug filer.

This introduces the problem of imbalanced output, but you can just
re-run the script repeatedly until you get a balanced-enough set of
lists. I think it should be possible to weight the random choice based
on the number of already-assigned bugs, but that seems better for a
followup.